requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
Flask>=2.2.0
google-generativeai>=0.3.0
